Pins are persisting!!!
TODO:
Deleted pins are still persisting. I need to associate the Pins with a
dictionary of the map annotations. Since I can’t use the annotations to
delete the Pin objects from the context, I need to use the annotation
to get the Pin, then use the context to delete that specific Pin object.