Almost working.
The problem, now, is that the memory address of the annotation view in
mapView:didSelectAnnotation/view() is not matching the key of the
dictionary where I’m storing the current pins (which, in turn, can
correspond to the Pin objects saved in the Core Data context). So, when
I click on a pin on the map, it isn’t returning a corresponding
annotation object from the current pins dictionary. I’m not quite sure
how to get around it yet.