Turns out, I need to take a new approach.
The original plan was to keep using all (up to) 250 photos returned by
a Flickr request, so that subsequent actions by the user to generate a
new album wouldn’t always need to contact Flickr for the next set of
photos. However, the 250 results just come back in the form of a
dictionary, from which the URLs need to be constructed and the Flickr
requests made, even if there are more unused entries in the dictionary.
SO - instead, I’m just gonna have it choose a random page and then take
the first 30 (or max number on the page) results.